move and rename storable.ts
